test(event,strategy): round the products this branch's expectations compute
Rebasing onto the widened fusible multiply-add guard flagged seven products in
this branch's own test fixtures, four of them in computed expected values:

  internal/strategy/delisting_test.go:642  wantEntryPrice
  internal/strategy/delisting_test.go:643  wantExitPrice
  internal/strategy/delisting_test.go:177  stopLevel
  internal/event/campaign_test.go:825      stopLevel

An expectation is exactly where this matters. A fused product in a test shifts
the expected value on arm64 by the same amount it shifts the production value,
so the test passes on both architectures while the production code disagrees
with itself -- which is why the guard's scope was widened past production
sources.

+++ internal/event/campaign_test.go
@@ -822,7 +822,7 @@ func validCampaignExitedByDelisting() event.CampaignExitedPayload {
 	exit := 179.5
 	n := proposalN
 	dpp := 1.0
-	stopLevel := entry - 2*n
+	stopLevel := entry - float64(2*n)
 	var unitQuantity int64 = 133
 	realisedResult := float64(133) * (exit - entry) * dpp
 	return event.CampaignExitedPayload{
